Add generic paged response
